+++ package.json
@@ -2,7 +2,7 @@
   "name": "material-table",
   "version": "1.35.2",
   "description": "Datatable for React based on https://material-ui.com/api/table/ with additional features",
-  "main": "dist/material-table.js",
+  "main": "dist/index.js",
   "types": "index.d.ts",
   "files": [
     "dist",
